requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
-e git+https://github.com/huggingface/datatrove.git@d8066af397c6458c6bc9deee45c664f0a247560f#egg=datatrove
dill==0.4.0
distro==1.9.0
exceptiongroup==1.2.2
fasteners==0.19
fasttext-wheel==0.9.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.2
idna==3.10
jieba==0.42.1
jiter==0.9.0
joblib==1.5.0
jsonlines==4.0.0
loguru==0.7.3
multiprocess==0.70.18
nltk==3.9.1
numpy>=1.23.5
openai==1.78.0
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pyarrow==20.0.0
pybind11==2.13.6
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
snownlp==0.12.3
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
win32_setctime==1.2.0
